fix add_type_restriction() implementation with IN() function (closes #138635)
When the variable type is defined by an IN() function, ``add_type_restriction``
removes all types in it that don't match the specified type. Since the list
of types is modified inplace, the iteration must be done on a copy, not on
the list itself.

There was a test for "add_type_restriction + IN()" case but since there
was only 2 types defined in the IN function, the problem was not exposed.