[stcheck] fix visit_constant checker: a etype inside a CAST is a good thing
Actually, if you put a CAST after a relation (!= ``is`` or ``is_instance_of``),
the checker raises an error saying that the etype (inside the CAST) must be
after a ``is`` or ``is_instance_of`` relation. In fact the CAST case was tested too late.

This allow to use CAST outside the selection section, so:
Closes #79232