[parsing] Raise BadRQLQuery if shortcut type settings is used without any variable
Before this 'Machin 12' is transformed to 'Any 12' since it expect a 'set type
shortcut', but since the selection doesn't hold any variable, the type
information is lost and no type checking is ever done.

We can not expect to have the schema at this point, which would allow to check
if the type exists. Even if we had, we would then need dedicated structures to
hold the information that we expect entity with eid '12' to be a 'Machin'...

Thus, it seems simpler to disallow usage of this simplified form without any
selected variable.

Closes #9234282.